Converted faux dashes to em-dashes
Ahhhhh, typography. My #1 pet peeve about the original text is the use
of hyphens where em-dashes belong. We can fuss over whether they should
be em-dashes or en-dashes, but I'm right and you're wrong, so what's
the use? Besides, the canon novels use em-dashes, so we should be
consistent. Yeah.